requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astronomer-cosmos
astronomer-cosmos[dbt-bigquery]
protobuf
apache-airflow-providers-google